[PLPR-8] - Changed Price and Discount fields type to int.

+++ internal/domain/ea-raw-item.go
@@ -3,8 +3,8 @@ package domain
 type EARawItem struct {
 	Name            string
 	Mnn             string
-	Price           float64
-	Discount        float64
+	Price           int
+	Discount        int
 	DiscountPercent int
 	Producer        string
 	Rating          float64
